# packages/shared/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/agents": {
      "post": {
        "operationId": "AgentsController_createAgent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Create a new agent",
        "tags": [
          "Agent"
        ]
      },
      "get": {
        "operationId": "AgentsController_listAgents",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "List agents with optional filtering and pagination",
        "tags": [
          "Agent"
        ]
      }
    },
    "/api/v1/agents/{slug}": {
      "get": {
        "operationId": "AgentsController_getAgentBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Agent slug",
            "examples": {
              "bySlug": {
                "summary": "By slug",
                "value": "alfred"
              }
            },
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Get an agent by slug",
        "tags": [
          "Agent"
        ]
      }
    },
    "/api/v1/agents/{actorId}": {
      "patch": {
        "operationId": "AgentsController_patchAgent",
        "parameters": [
          {
            "name": "actorId",
            "required": true,
            "in": "path",
            "description": "Agent actor ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Patch an agent (update system prompt and/or status triggers)",
        "tags": [
          "Agent"
        ]
      }
    },
...
  },
  "info": {
    "title": "Traff API",
    "description": "The Traff API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-Cookie": {
        "type": "http",
        "in": "cookie",
        "name": "access_token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateAgentDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Unique, human-readable identifier for the agent",
            "example": "buddy"
          },
          "name": {
            "type": "string",
            "description": "Display name for the agent",
            "example": "Buddy"
          },
          "type": {
            "type": "string",
            "description": "Type of agent (provider)",
            "example": "claude",
            "enum": [
              "claude",
              "codex",
              "opencode",
              "other"
            ],
            "default": "claude"
          },
          "description": {
            "type": "string",
            "description": "Short description of what this agent does",
            "example": "A helpful assistant agent"
          },
          "systemPrompt": {
            "type": "string",
            "description": "Core instructions/persona for this agent",
            "example": "You are a helpful assistant that helps users with tasks."
          },
          "statusTriggers": {
            "type": "array",
            "description": "Task statuses that will trigger this agent to activate",
            "example": [
              "IN_PROGRESS",
              "FOR_REVIEW"
            ],
            "default": [],
            "items": {
              "type": "string",
              "enum": [
                "NOT_STARTED",
                "IN_PROGRESS",
                "FOR_REVIEW",
                "DONE"
              ]
            }
          },
          "allowedTools": {
            "description": "List of tool identifiers this agent is allowed to use",
            "example": [
              "tasks.createTask",
              "tasks.readTask",
              "context.search"
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether this agent is available for assignment",
            "example": true,
            "default": true
          },
          "concurrencyLimit": {
            "type": "number",
            "description": "Max number of tasks this agent can process in parallel",
            "example": 5
          }
        },
        "required": [
          "slug",
          "name",
          "systemPrompt"
        ]
      },
      "AgentResponseDto": {
        "type": "object",
        "properties": {
          "actorId": {
            "type": "string",
            "description": "Unique identifier for the actor representing this agent",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "slug": {
            "type": "string",
            "description": "Unique, human-readable identifier",
            "example": "buddy"
          },
          "name": {
            "type": "string",
            "description": "Display name for the agent",
            "example": "Buddy"
          },
          "type": {
            "type": "string",
            "description": "Type of agent (provider)",
            "example": "claude",
            "enum": [
              "claude",
              "codex",
              "opencode",
              "other"
            ]
          },
          "description": {
            "type": "object",
            "description": "Short description of what this agent does",
            "example": "A helpful assistant agent"
          },
          "systemPrompt": {
            "type": "string",
            "description": "Core instructions/persona for this agent",
            "example": "You are a helpful assistant that helps users with tasks."
          },
          "statusTriggers": {
            "description": "List of status that trigger this agent",
            "example": [
              "NOT_STARTED",
              "IN_PROGRESS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedTools": {
            "description": "List of tool identifiers this agent is allowed to use",
            "example": [
              "tasks.createTask",
              "tasks.readTask",
              "context.search"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether this agent is available for assignment",
            "example": true
          },
          "concurrencyLimit": {
            "type": "object",
            "description": "Max number of tasks this agent can process in parallel",
            "example": 5
          },
          "rowVersion": {
            "type": "number",
            "description": "Row version for optimistic locking",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "description": "Agent creation timestamp",
            "example": "2025-11-28T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Agent last update timestamp",
            "example": "2025-11-28T10:30:00.000Z"
          },
          "deletedAt": {
            "type": "object",
            "description": "Agent deletion timestamp (soft delete)",
            "example": null
          }
        },
        "required": [
          "actorId",
          "slug",
          "name",
          "type",
          "systemPrompt",
          "statusTriggers",
          "allowedTools",
          "isActive",
          "rowVersion",
          "createdAt",
          "updatedAt"
        ]
      },
      "AgentListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "List of agents",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of agents matching the query",
            "example": 42
          },
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of items per page",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages",
            "example": 3
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PatchAgentDto": {
        "type": "object",
        "properties": {
          "systemPrompt": {
            "type": "string",
            "description": "Core instructions/persona for this agent",
            "example": "You are a helpful assistant that helps users with tasks."
          },
          "statusTriggers": {
            "type": "array",
            "description": "Task statuses that will trigger this agent to activate",
            "example": [
              "IN_PROGRESS",
              "FOR_REVIEW"
            ],
            "items": {
              "type": "string",
              "enum": [
                "NOT_STARTED",
                "IN_PROGRESS",
                "FOR_REVIEW",
                "DONE"
              ]
            }
          }
        }
      },
...
    }
  }
}